fixes for actually using this in a product
Support for different handlers at the same depth

This fixes an issue where the RouteHandler would not invoke the
callbacks properly if the current handler changes but the depth does
not. The hierarchy we're seeing this issue in is like this:

- Root
  - AuthorCourse
    - Summary (<DefaultRoute />)
    - Learners (<Route />)

When navigating from Summary => Learners, I expect:

  1. Summary#teardown to be called
  2. Learners#setup to be called

This patch makes that happen.

Also:

- No longer ignoring modules/ in NPM (0.1.8)
- Use native Promise
- Pass in the loading promise for error handling